New manage/update
Now using an updated manage/update script utilizing the environment.yaml file and Python 3.9. Also added troubleshooting section on how to remove the hazus_env. Removed preprocessing launcher script and mention of it in the help docs and renamed functions scripts. Removed old help docs. New Hazus icon.
