+++ .gitignore
@@ -1,20 +1,108 @@
-# ignore files
-*.ipynb
-
-# ignore folders
-.ipynb_checkpoints/
-.vscode/
+# Byte-compiled / optimized / DLL files
 __pycache__/
+*.py[cod]
+*$py.class
+
+# C extensions
+*.so
+
+# Distribution / packaging
+.Python
+build/
+develop-eggs/
+dist/
+downloads/
+eggs/
+.eggs/
+lib/
+lib64/
+parts/
+sdist/
+var/
+wheels/
+*.egg-info/
+.installed.cfg
+*.egg
+MANIFEST
+
+# PyInstaller
+#  Usually these files are written by a python script from a template
+#  before PyInstaller builds the exe, so as to inject date/other infos into it.
+*.manifest
+*.spec
+
+# Installer logs
+pip-log.txt
+pip-delete-this-directory.txt
+
+# Unit test / coverage reports
+htmlcov/
+.tox/
+.coverage
+.coverage.*
+.cache
+nosetests.xml
+coverage.xml
+*.cover
+.hypothesis/
+.pytest_cache/
+
+# Translations
+*.mo
+*.pot
+
+# Django stuff:
+*.log
+local_settings.py
+db.sqlite3
+
+# Flask stuff:
+instance/
+.webassets-cache
+
+# Scrapy stuff:
+.scrapy
+
+# Sphinx documentation
+docs/_build/
+
+# PyBuilder
+target/
+
+# Jupyter Notebook
+.ipynb_checkpoints
+
+# pyenv
+.python-version
+
+# celery beat schedule file
+celerybeat-schedule
+
+# SageMath parsed files
+*.sage.py
+
+# Environments
+.env
+.venv
+env/
+venv/
+ENV/
+env.bak/
+venv.bak/
+
+# Spyder project settings
+.spyderproject
+.spyproject
+
+# Rope project settings
+.ropeproject
+
+# mkdocs documentation
+/site
+
+# mypy
+.mypy_cache/
+
+#custom folders
 _Python_env
-GDAL/
-UDF/HI_Honolulu_UDF_TestErrors.csv
-rasters/HI_Honolulu_GAT.ovr
-rasters/HI_Honolulu_GAT.tfw
-rasters/HI_Honolulu_GAT.tif
-rasters/HI_Honolulu_GAT.tif.aux.xml
-rasters/HI_Honolulu_GAT.tif.ovr
-rasters/HI_Honolulu_GAT.xml
-UDF/HI_Honolulu_UDF_HI_Honolulu_GAT.csv
-UDF/HI_Honolulu_UDF_HI_Honolulu_GAT_sorted.csv
-rasters/__init__.py
-*.pyc
+GDAL
