+++ src/__init__.py
@@ -10,4 +10,4 @@
 
 """
 
-__version__ = '0.0.7'
+__version__ = '0.0.6'
