+++ src/environment.yaml
@@ -1,0 +1,162 @@
+name: hazus_env
+channels:
+  - nhrap
+  - nsls2forge
+  - conda-forge
+  - anaconda
+  - defaults
+dependencies:
+  - affine=2.3.0=py_0
+  - arabic_reshaper=2.1.1=pyhe636693_0
+  - asn1crypto=1.4.0=pyh9f0ad1d_0
+  - attrs=20.3.0=pyhd3deb0d_0
+  - bcrypt=3.2.0=py39hb82d6ee_1
+  - blas=1.0=mkl
+  - brotlipy=0.7.0=py39hb82d6ee_1001
+  - bzip2=1.0.8=h8ffe710_4
+  - ca-certificates=2020.12.5=h5b45459_0
+  - certifi=2020.12.5=py39hcbf5309_1
+  - cffi=1.14.4=py39h0878f49_1
+  - cfitsio=3.470=h0af3d06_7
+  - chardet=4.0.0=py39hcbf5309_1
+  - click=7.1.2=pyh9f0ad1d_0
+  - click-plugins=1.1.1=py_0
+  - cligj=0.7.1=pyhd8ed1ab_0
+  - configparser=5.0.1=py_0
+  - cryptography=3.3.1=py39h0878f49_1
+  - curl=7.71.1=h4b64cdc_8
+  - cycler=0.10.0=py_2
+  - decorator=4.4.2=py_0
+  - deprecated=1.2.10=pyh9f0ad1d_0
+  - descartes=1.1.0=py_4
+  - expat=2.2.10=h39d44d4_0
+  - fiona=1.8.17=py39h6119223_1
+  - freetype=2.10.4=h546665d_1
+  - freexl=1.0.5=hd288d7e_1002
+  - future=0.18.2=py39hcbf5309_3
+  - gdal=3.1.4=py39hda8168b_0
+  - geopandas=0.8.2=pyhd8ed1ab_0
+  - geos=3.8.1=he025d50_0
+  - geotiff=1.6.0=h8884d1a_3
+  - gettext=0.19.8.1=hb01d8f6_1002
+  - glib=2.65.0=he4de6d7_0
+  - hazpy=0.0.1=py39_0
+  - hdf4=4.2.13=h0e5069d_1004
+  - hdf5=1.10.6=nompi_h5268f04_1114
+  - html5lib=1.1=pyh9f0ad1d_0
+  - icu=67.1=h33f27b4_0
+  - idna=2.10=pyh9f0ad1d_0
+  - intel-openmp=2020.3=h57928b3_311
+  - jenkspy=0.1.4=py39h4cdbadb_1002
+  - joblib=1.0.0=pyhd8ed1ab_0
+  - jpeg=9d=h8ffe710_0
+  - kealib=1.4.14=ha3510f1_0
+  - kiwisolver=1.3.1=py39h2e07f2f_1
+  - krb5=1.17.2=hbae68bd_0
+  - libblas=3.9.0=8_mkl
+  - libboost=1.65.1=vc14h7b94727_3
+  - libcblas=3.9.0=8_mkl
+  - libclang=10.0.1=default_hf44288c_1
+  - libcurl=7.71.1=h4b64cdc_8
+  - libffi=3.2.1=ha925a31_1007
+  - libgdal=3.1.4=h0e5aa5a_0
+  - libiconv=1.15=vc14h29686d3_5
+  - libkml=1.3.0=hc65d273_3
+  - liblapack=3.9.0=8_mkl
+  - libnetcdf=4.7.4=nompi_h3a9aa94_107
+  - libpng=1.6.37=h1d00b33_2
+  - libpq=12.3=h4f54205_3
+  - libspatialindex=1.9.3=he025d50_3
+  - libspatialite=5.0.0=h37d8b57_0
+  - libssh2=1.9.0=hb06d900_5
+  - libtiff=4.2.0=hc10be44_0
+  - libwebp-base=1.2.0=h8ffe710_0
+  - libxml2=2.9.10=h5d81f1c_2
+  - lz4-c=1.9.3=h8ffe710_0
+  - m2w64-gcc-libgfortran=5.3.0=6
+  - m2w64-gcc-libs=5.3.0=7
+  - m2w64-gcc-libs-core=5.3.0=7
+  - m2w64-gmp=6.1.0=2
+  - m2w64-libwinpthread-git=5.0.0.4634.697f757=2
+  - mapclassify=2.4.2=pyhd8ed1ab_0
+  - matplotlib=3.3.2=py39hcbf5309_1
+  - matplotlib-base=3.3.2=py39h62a766e_1
+  - mkl=2020.4=hb70f87d_311
+  - mkl-service=2.3.0=py39h4cdbadb_2
+  - mkl_fft=1.2.0=py39h8512cfa_1
+  - mkl_random=1.0.2=py39h848d8c7_0
+  - msys2-conda-epoch=20160418=1
+  - munch=2.5.0=py_0
+  - networkx=2.5=py_0
+  - numpy=1.19.2=py39h729668d_0
+  - numpy-base=1.19.2=py39hbd0edd7_0
+  - olefile=0.46=pyh9f0ad1d_1
+  - openjpeg=2.3.1=h48faf41_3
+  - openssl=1.1.1i=h8ffe710_0
+  - palettable=3.3.0=py_0
+  - pandas=1.1.4=py39h2e25243_0
+  - paramiko=2.7.2=pyh9f0ad1d_0
+  - patsy=0.5.1=py_0
+  - pcre=8.44=ha925a31_0
+  - pillow=8.1.0=py39hbc3ea3a_1
+  - pip=21.0.1=pyhd8ed1ab_0
+  - poppler=0.89.0=h0cd1227_0
+  - poppler-data=0.4.10=0
+  - postgresql=12.3=h0f1a9bc_3
+  - proj=7.1.1=h7d85306_3
+  - psycopg2=2.8.6=py39h0878f49_1
+  - pycparser=2.20=pyh9f0ad1d_2
+  - pynacl=1.4.0=py39hb3671d1_2
+  - pyodbc=4.0.30=py39h415ef7b_1
+  - pyopenssl=20.0.1=pyhd8ed1ab_0
+  - pyparsing=2.4.7=pyh9f0ad1d_0
+  - pypdf2=1.26.0=py_0
+  - pyproj=2.6.1.post1=py39hd2b018c_3
+  - pyqt=5.12.3=py39hcbf5309_7
+  - pyqt-impl=5.12.3=py39h415ef7b_7
+  - pyqt5-sip=4.19.18=py39h415ef7b_7
+  - pyqtchart=5.12=py39h415ef7b_7
+  - pyqtwebengine=5.12.1=py39h415ef7b_7
+  - pysal=2.1.0=py_0
+  - pysocks=1.7.1=py39hcbf5309_3
+  - python=3.9.1=h7840368_5_cpython
+  - python-bidi=0.4.2=py_0
+  - python-dateutil=2.8.1=py_0
+  - python_abi=3.9=1_cp39
+  - pytz=2021.1=pyhd8ed1ab_0
+  - qt=5.12.9=hb2cf2c5_0
+  - rasterio=1.1.8=py39hfec4536_0
+  - reportlab=3.5.60=py39h12d40c5_0
+  - requests=2.25.1=pyhd3deb0d_0
+  - rtree=0.9.7=py39h09fdee3_1
+  - scikit-learn=0.23.2=py39he931e04_2
+  - scipy=1.5.3=py39hc0c34ad_0
+  - seaborn=0.11.1=hd8ed1ab_1
+  - seaborn-base=0.11.1=pyhd8ed1ab_1
+  - setuptools=49.6.0=py39hcbf5309_3
+  - shapely=1.7.1=py39hdfdf89f_1
+  - six=1.15.0=pyh9f0ad1d_0
+  - snuggs=1.4.7=py_0
+  - sqlalchemy=1.3.23=py39hb82d6ee_0
+  - sqlite=3.34.0=h8ffe710_0
+  - statsmodels=0.12.1=py39h5d4886f_1
+  - threadpoolctl=2.1.0=pyh5ca1d4c_0
+  - tiledb=2.1.6=hf84e3da_1
+  - tk=8.6.10=h8ffe710_1
+  - tornado=6.1=py39hb82d6ee_1
+  - tqdm=4.56.0=pyhd8ed1ab_0
+  - tzdata=2021a=he74cb21_0
+  - urllib3=1.26.3=pyhd8ed1ab_0
+  - utm=0.6.0=pyh9f0ad1d_0
+  - vc=14.2=hb210afc_2
+  - vs2015_runtime=14.28.29325=h5e1d092_0
+  - webencodings=0.5.1=py_1
+  - wheel=0.36.2=pyhd3deb0d_0
+  - win_inet_pton=1.1.0=py39hcbf5309_2
+  - wincertstore=0.2=py39hcbf5309_1006
+  - wrapt=1.12.1=py39hb82d6ee_3
+  - xerces-c=3.2.3=h0e60522_2
+  - xhtml2pdf=0.2.5=py_0
+  - xz=5.2.5=h62dcd97_1
+  - zlib=1.2.11=vc14h1cdd9ab_1
+  - zstd=1.4.8=h4e2f164_1
