+++ src/update.py
@@ -1,11 +1,14 @@
 try:
-    from manage import internetConnected, checkForHazPyUpdates, checkForToolUpdates
-    checkForHazPyUpdates()
-    print(3)
-    checkForToolUpdates()
-
-except:
+    try:
+        from src.manage import Manage
+        manage = Manage()
+    except:
+        from manage import Manage
+        manage = Manage()
+    manage.checkForToolUpdates()
+    manage.checkForHazPyUpdates()
+except Exception as e:
     import ctypes
     import sys
     messageBox = ctypes.windll.user32.MessageBoxW
-    messageBox(0, "Unexpected error:" + sys.exc_info()[0] + " | If this problem persists, contact hazus-support@riskmapcds.com.", "HazPy", 0x1000)
+    messageBox(0, u"Unexpected error:" + str(sys.exc_info()[0]) + u" | If this problem persists, contact hazus-support@riskmapcds.com.", u"HazPy", 0x1000)
